backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Marshmallow==3.20.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==7.4.0
pytest-flask==1.2.0
